Enable MacOS tests (#1352)
