+++ .azure-pipelines/stage-e2e.yml
@@ -18,6 +18,12 @@ stages:
         Py37-Win64:
           python.version: '3.7'
           image: 'vs2017-win2016'
+        Py36-Mac:
+          python.version: '3.6'
+          image: 'macos-latest'
+        Py37-Mac:
+          python.version: '3.7'
+          image: 'macos-latest'
 
     pool:
       vmImage: $[ variables['image'] ]
@@ -47,7 +53,7 @@ stages:
       displayName: 'Run e2e'
       env:
         CLIENT_TEST_E2E_USER_NAME: $(e2e.token)
-        PYTEST_XDIST_NUM_THREADS: 16
+        PYTEST_XDIST_NUM_THREADS: 8
 
     - script: |
         pip install codecov
